Create Computer-Object_Tool.ps1
This is a tool I wrote to scan for computer objects that have not updated the PasswordLastSet attribute and use that information to disable unused accounts.
Not a perfect solution but a good start for a general frame work. Does work well for most workstation objects. Care should be taken when running against servers as some objects like cluster server objects will not update this attribute.